% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_log_ratios}
\alias{compute_log_ratios}
\title{Compute Log Ratios}
\usage{
compute_log_ratios(reference, sample, mat)
}
\arguments{
\item{reference}{Column name or integer index of the reference.}

\item{sample}{Column name or integer index of the sample.}

\item{mat}{Numeric matrix or sparse Matrix::dgCMatrix.}
}
\value{
Numeric vector of finite log-ratios.
}
\description{
Compute log(reference/sample) for a given pair of columns, efficiently for dgCMatrix.
}
