% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_scaling_factor}
\alias{compute_scaling_factor}
\title{Compute Scaling Factor}
\usage{
compute_scaling_factor(mat, reference_name, sample_name, model)
}
\arguments{
\item{mat}{Matrix or sparse matrix of values}

\item{reference_name}{Column name of the reference}

\item{sample_name}{Column name of the sample}

\item{model}{Model object with interval}
}
\value{
Numeric scaling factor (sample to reference)
}
\description{
Computes the pseudo-coverage ratio between a reference and sample using model cutoffs to define valid log-ratio ranges.
}
