% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{convert_to_numeric}
\usage{
convert_to_numeric(x)
}
\arguments{
\item{x}{A matrix or data frame.}
}
\value{
A numeric matrix or sparse dgCMatrix.
}
\description{
Converts input to numeric format, preserving sparse matrix structure.
}
