% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_sparse_matrix_optimized}
\alias{load_sparse_matrix_optimized}
\title{load_sparse_matrix_optimized}
\usage{
load_sparse_matrix_optimized(file_path, row_name_index = 1, sep = "\\t")
}
\arguments{
\item{file_path}{Path to the input matrix file.}

\item{row_name_index}{Column index for rownames. Default = 1.}

\item{sep}{Field separator (default tab).}
}
\value{
A dgCMatrix with rownames set.
}
\description{
Fast loader that returns a sparse dgCMatrix from a delimited text file.
}
