% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{median_of_ratios_normalization}
\alias{median_of_ratios_normalization}
\title{median_of_ratios_normalization}
\usage{
median_of_ratios_normalization(data)
}
\arguments{
\item{data}{data frame to apply quantile norm}
}
\value{
a data frame  normalized
}
\description{
median_of_ratios_normalization
}
