% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{noceros}
\alias{noceros}
\title{noceros To count number of non-zero elements}
\usage{
noceros(x, num = TRUE, k = 0)
}
\arguments{
\item{x}{vector to check zeros}

\item{num}{if vector is numeric TRUE FALSE}

\item{k}{default is 0}
}
\value{
return non zero elements
}
\description{
noceros To count number of non-zero elements
}
