% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_design}
\alias{parse_design}
\title{parse_design}
\usage{
parse_design(design_path)
}
\arguments{
\item{design_path}{File path or in-memory data.table.}
}
\value{
A validated design data.table with Sample_ID as rownames.
}
\description{
Parses a design matrix from a file or data.table.
}
