% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_input_matrix}
\alias{parse_input_matrix}
\title{parse_input_matrix}
\usage{
parse_input_matrix(mat_path, row_name_index = 1)
}
\arguments{
\item{mat_path}{File path or matrix-like object.}

\item{row_name_index}{Column index for rownames (only used if \code{mat_path} is a file).}
}
\value{
A dgCMatrix.
}
\description{
Parses and converts a file path or R object into a sparse dgCMatrix.
}
