% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pair_model.R
\name{plot_pair_model}
\alias{plot_pair_model}
\title{plot_pair_model Plots fitted distributions to the ratios.}
\usage{
plot_pair_model(ll)
}
\arguments{
\item{ll}{List composed by: "ratio" for the selected pair, "n_pop" i.e. n° of expected populations,
"sigma_times" n° of sigmas to select invariant observations, "dist_family" distribution famly see mixsmsn, ...}
}
\value{
Returns automatically pdf saved within "saving_path"/Norm_pair_plots/ of the fitting and QC
}
\description{
plot_pair_model Plots fitted distributions to the ratios.
}
\examples{
\dontrun{
# Typical usage
ll <- list(
  "ratio" = ratio, "n_pop" = n_pop_reference,
  "sigma_times" = sigma_times, "dist_family" = dist_family, "index" = x
)
plot_pair_model(ll)
}

}
