% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_normalisation}
\alias{quantile_normalisation}
\title{quantile_normalisation}
\usage{
quantile_normalisation(df)
}
\arguments{
\item{df}{data frame to apply quantile norm}
}
\value{
a data frame  normalized
}
\description{
quantile_normalisation
}
