% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uqua}
\alias{uqua}
\title{Upper quartile normalisation}
\usage{
uqua(datos, long = 1000, lc = 1, k = 0)
}
\arguments{
\item{datos}{Matrix containing the read counts for each sample.}

\item{long}{Numeric vector containing the length of the features. If long == 1000, no length correction is applied (no matter the value of parameter lc).}

\item{lc}{Correction factor for length normalization. This correction is done by dividing the counts vector by (length/1000)^lc.
If lc = 0, no length correction is applied. By default, lc = 1 for RPKM and lc = 0 for the other methods.}

\item{k}{Counts equal to 0 are changed to k in order to avoid indeterminations when applying logarithms, for instance. By default, k = 0.}
}
\value{
a data frame  normalized
}
\description{
Upper quartile normalisation
}
