% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VarianceStabilized}
\alias{VarianceStabilized}
\title{VarianceStabilized}
\usage{
VarianceStabilized(norm_mat)
}
\arguments{
\item{norm_mat}{A numeric matrix to be variance-stabilized.}
}
\value{
A variance-stabilized matrix.
}
\description{
Applies a variance stabilization transformation to a normalized matrix.
}
