% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{median_of_ratios_normalization}
\alias{median_of_ratios_normalization}
\title{median_of_ratios_normalization}
\usage{
median_of_ratios_normalization(data)
}
\arguments{
\item{data}{A data frame or matrix of counts.}
}
\value{
A normalized data frame.
}
\description{
Median-of-ratios normalization method (as in DESeq2).
}
