% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{noceros}
\alias{noceros}
\title{noceros}
\usage{
noceros(x, num = TRUE, k = 0)
}
\arguments{
\item{x}{Numeric vector.}

\item{num}{Logical. If \code{TRUE}, returns the count; if \code{FALSE}, returns indices (default = TRUE).}

\item{k}{Threshold. Elements greater than \code{k} are considered non-zero (default = 0).}
}
\value{
Either a count or vector of indices of non-zero elements.
}
\description{
Count or retrieve non-zero elements from a numeric vector.
}
