% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_fit.R
\name{pair_fit}
\alias{pair_fit}
\title{pair_fit execute pair wise fitting of Skewed normal on the freq. distribution of the ratios.}
\usage{
pair_fit(ll)
}
\arguments{
\item{ll}{List composed by: "ratio" for the selected pair, "n_pop" i.e. n° of expected populations,
"sigma_times" n° of sigmas to select invariant observations, "dist_family" distribution famly see mixsmsn, ...}
}
\value{
Return EM algorithm output for FM-SMSN distributions with additional fields: n_pop,means,modes,fre  q,freq_dist,sigma_norm
}
\description{
pair_fit execute pair wise fitting of Skewed normal on the freq. distribution of the ratios.
}
\examples{
\dontrun{
# Typical usage
ll <- list(
  "ratio" = ratio, "n_pop" = n_pop_reference,
  "sigma_times" = sigma_times, "dist_family" = dist_family, "index" = x
)
model <- pair_fit(ll)
}

}
