% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot.R
\name{QC_plot}
\alias{QC_plot}
\title{QC_plot Function performing QC by PCA analysis.}
\usage{
QC_plot(
  gnmsn_obj,
  design_path,
  custom_table = NULL,
  saving_path = NULL,
  norm_methods = c("quantile_normalisation", "median_of_ratios_normalization", "uqua",
    "tmm")
)
}
\arguments{
\item{gnmsn_obj}{path to the matrix (n * m having features i.e. genes, proteins, chipseq peaks in the raws and individual samples in the columns)}

\item{design_path}{design matrix describing the samples. design matrix must be composed by:  Sample_ID, Sample_Condition and Sample_Replicate columns (layers in Sample_conditions will be defined by "_" separator)}

\item{custom_table}{vector of Sample_ID names to be used as initial reference for normalisation. If random specified it select randomly 30\% of the samples (To be improoved considering global coverage and intensity in order to avoid failed samples) (default NULL)}

\item{saving_path}{full path where results are saved . default to the working directory}

\item{norm_methods}{normalisation methods to compare the GNMSN : can be one of "quantile_normalisation", "median_of_ratios_normalization", "uqua", "tmm"}
}
\value{
Save plots and table to the "saving_pth"/Norm_quality/
}
\description{
QC_plot Function performing QC by PCA analysis.
}
\examples{
\dontrun{
# Typical usage
param <- BiocParallel::MulticoreParam(workers = 2, progressbar = TRUE)

Result <- RunNorm(mat,
  deg_design,
  fix_reference = "random",
  row_name_index = 1,
  saving_path = out_dir,
  n_pop = 1, BiocParam = param
)

QC_plot(Result, deg_design, saving_path = out_dir)
}

}
