% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_normalisation}
\alias{quantile_normalisation}
\title{quantile_normalisation}
\usage{
quantile_normalisation(df)
}
\arguments{
\item{df}{A numeric data frame or matrix to normalize.}
}
\value{
A quantile-normalized data frame.
}
\description{
Apply quantile normalization to a data frame.
}
