% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNorm.R
\name{RunNorm}
\alias{RunNorm}
\title{RunNorm function to execute the whole normalisation flow.}
\usage{
RunNorm(
  mat_path,
  design_path,
  fix_reference = NULL,
  row_name_index = 1,
  n_pop = 1,
  n_pop_reference = 1,
  saving_path = NULL,
  sigma_times = 1,
  dist_family = "Skew.normal",
  Norm_plot = TRUE,
  Save_results = TRUE,
  BiocParam = NULL
)
}
\arguments{
\item{mat_path}{path to the matrix (n * m having features i.e. genes, proteins, chipseq peaks in the raws and individual samples in the columns)}

\item{design_path}{design matrix describing the samples. design matrix must be composed by:  Sample_ID, Sample_Condition and Sample_Replicate columns (layers in Sample_conditions will be defined by "_" separator)}

\item{fix_reference}{vector of Sample_ID names to be used as initial reference for normalisation. If random specified it select randomly 30\% of the samples (To be improoved considering global coverage and intensity in order to avoid failed samples) (default NULL)}

\item{row_name_index}{column index where to find the rownames of the matrix (default 1)}

\item{n_pop}{n° of populations expected (default NULL)}

\item{n_pop_reference}{n° of expected populations in the initial within references normalisation (default NULL)}

\item{saving_path}{full path where results are saved default to the working directory}

\item{sigma_times}{number of sigmas defining lower and upper bounds to include features (i.e. genes, peaks, proteins ,..) considered to be unchanged relative to the mean distribution}

\item{dist_family}{distribution family see package mixsmsn (default "Skew.normal")}

\item{Norm_plot}{Whether or not to save individual plots during normalisation (defulat TRUE)}

\item{Save_results}{whether to save individual plots and results (default TRUE)}

\item{BiocParam}{BiocParallel param object to be specified as : "param <- BiocParallel::MulticoreParam(workers=2,progressbar = TRUE)" (default NULL)}
}
\value{
Return a list which includes the fields: scaling_factors; ori_mat; norm_mat and model_list
}
\description{
RunNorm function to execute the whole normalisation flow.
}
\examples{
\dontrun{
# Typical usage
param <- BiocParallel::MulticoreParam(workers = 2, progressbar = TRUE)

Result <- RunNorm(mat,
  deg_design,
  fix_reference = "random",
  row_name_index = 1,
  saving_path = out_dir,
  n_pop = 1, BiocParam = param
)
}

}
