% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmm}
\alias{tmm}
\title{tmm: Trimmed Mean of M values normalization (Robinson & Oshlack, 2010)}
\usage{
tmm(
  datos,
  long = 1000,
  lc = 1,
  k = 0,
  refColumn = NULL,
  logratioTrim = 0.3,
  sumTrim = 0.05,
  doWeighting = TRUE,
  Acutoff = -1e+10
)
}
\arguments{
\item{datos}{Matrix containing the read counts for each sample.}

\item{long}{Numeric vector containing the length of the features. If long == 1000, no length correction is applied (no matter the value of parameter lc).}

\item{lc}{Correction factor for length normalization. This correction is done by dividing the counts vector by (length/1000)^lc.
If lc = 0, no length correction is applied. By default, lc = 1 for RPKM and lc = 0 for the other methods.}

\item{k}{Counts equal to 0 are changed to k in order to avoid indeterminations when applying logarithms, for instance. By default, k = 0.}

\item{refColumn}{Column to use as reference (only needed for tmm function).}

\item{logratioTrim}{Amount of trim to use on log-ratios ("M" values) (only needed for tmm function).}

\item{sumTrim}{Amount of trim to use on the combined absolute levels ("A" values) (only needed for tmm function).}

\item{doWeighting}{Logical, whether to compute (asymptotic binomial precision) weights (only needed for tmm function).}

\item{Acutoff}{Cutoff on "A" values to use before trimming (only needed for tmm function).}
}
\value{
a data frame  normalized
}
\description{
tmm: Trimmed Mean of M values normalization (Robinson & Oshlack, 2010)
}
