% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uqua}
\alias{uqua}
\title{uqua}
\usage{
uqua(datos, long = 1000, lc = 1, k = 0)
}
\arguments{
\item{datos}{Numeric matrix of read counts.}

\item{long}{Numeric vector of feature lengths. If 1000, no correction is applied.}

\item{lc}{Length correction exponent. Default = 1.}

\item{k}{Replacement for zeros to avoid log issues. Default = 0.}
}
\value{
A normalized numeric data frame.
}
\description{
Upper-quartile normalization with optional length correction.
}
