% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VarianceStabilized}
\alias{VarianceStabilized}
\title{Variance stabilisation}
\usage{
VarianceStabilized(norm_mat)
}
\arguments{
\item{norm_mat}{data frame to apply Variance stabilisation}
}
\value{
Variance stabilized data frame
}
\description{
Variance stabilisation
}
